Move shell menu and hotkey to ui/ subdir
